+++ src/main/java/dev/fulmineo/companion_bats/CompanionBats.java
@@ -93,6 +93,7 @@ public class CompanionBats implements ModInitializer {
 	public static final Item WOLF_PELT = new CompanionBatAccessoryItem("wolf_pelt", CompanionBatAbility.ATTACK_PASSIVE, 1, new FabricItemSettings().group(GROUP).maxCount(1));
 	public static final Item PLATED_BOOTS = new CompanionBatAccessoryItem("plated_boots", CompanionBatAbility.INCREASED_ARMOR, 1, new FabricItemSettings().group(GROUP).maxCount(1));
 	public static final Item PUMPKIN_BANDANA = new CompanionBatAccessoryItem("pumpkin_bandana", CompanionBatAbility.NATURAL_REGENERATION, 1, new FabricItemSettings().group(GROUP).maxCount(1));
+	public static final Item ANKH = new CompanionBatAccessoryItem("ankh", CompanionBatAbility.MAGIC_PROTECTION, 1, new FabricItemSettings().group(GROUP).maxCount(1));
 
 	// Armors
 
@@ -137,6 +138,7 @@ public class CompanionBats implements ModInitializer {
 		Registry.register(Registry.ITEM, new Identifier(MOD_ID, "wolf_pelt"), 		  	WOLF_PELT);
 		Registry.register(Registry.ITEM, new Identifier(MOD_ID, "plated_boots"), 		PLATED_BOOTS);
 		Registry.register(Registry.ITEM, new Identifier(MOD_ID, "pumpkin_bandana"), 	PUMPKIN_BANDANA);
+		Registry.register(Registry.ITEM, new Identifier(MOD_ID, "ankh"), 				ANKH);
 
 		// Armors